requirements.txt
openai>=1.75.0
python-dotenv>=1.0.1
requests>=2.32.3
yt-dlp>=2025.3.31
